Issue #17 - Move provider helper from utils to his own folder
